#48 Permission for users on backend

+++ app/Http/Controllers/BranchController.php
@@ -55,6 +55,8 @@ class BranchController extends Controller
      */
     public function edit(Branch $branch)
     {
+        $this->authorize('update', $branch);
+
         $sections = Section::all();
         $sections = SectionResource::collection($sections)->resolve();
 
@@ -68,6 +70,8 @@ class BranchController extends Controller
      */
     public function update(UpdateRequest $request, Branch $branch)
     {
+        $this->authorize('update', $branch);
+
         $data = $request->validated();
         $branch->update($data);
 
@@ -79,6 +83,8 @@ class BranchController extends Controller
      */
     public function destroy(Branch $branch)
     {
+        $this->authorize('delete', $branch);
+
         $branch->delete();
         return redirect()->route('sections.index');
     }
